% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.mudens.R
\name{lines.mudens}
\alias{lines.mudens}
\title{Plots estimated density function from an object of class 'mudens'.}
\usage{
\method{lines}{mudens}(x,...)
}
\arguments{
\item{x}{Object of class mudens (output from calling mudens function)}

\item{...}{Additional arguments for basic \code{lines(.)} function to be passed along.}
}
\description{
Add connected segments to a plot of estimated density function
             from an object of class 'mudens'.
}
\details{
Plots estimated density function from an object of class 'mudens'.

Plots estimated density function from an object of class 'mudens'.
Default time limits are those provided to mudens,
          which default to zero and the max observed time.
          Default y-axis limits are 0 and the maximum estimated density.
}
\references{
Hess, K.R. and Zhong, M. Density Function Estimation for Possibly Right-Censored Data Using Kernel Functions. Submitted.
}
\seealso{
\link{mudens}
}
\author{
Kenneth R. Hess
}
