% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mudens.R
\name{plot.mudens}
\alias{plot.mudens}
\title{Plots estimated density function from an object of class 'mudens'.}
\usage{
\method{plot}{mudens}(x,...)
}
\arguments{
\item{x}{Object of class mudens (output from calling mudens function)}

\item{...}{Additional arguments of basic \code{plot(.)} function to be passed along.}
}
\description{
Plot estimated density function from an object of class 'mudens'.
}
\details{
Plots estimated density function from an object of class 'mudens'.

Default time limits are those provided to mudens,
which default to zero and the time corresponding to
the maximum value of the times. Default y-axis limits are 0 and
the maximum estimated density.
}
\references{
Hess, K.R. and Zhong, M. Density Function Estimation for Possibly Right-Censored Data Using Kernel Functions. Submitted.
}
\seealso{
\link{mudens}
}
\author{
Kenneth R. Hess
}
